refactor: move CustomPool to submodule
we're going to store all the database logic in src/bot/database. also
flesh out the CustomPool object to add event handlers.
